+++ packages/rum/test/index.ts
@@ -1,2 +1,3 @@
 export * from './mockWorker'
 export * from './utils'
+export * from './mutationPayloadValidator'
